Update README.md
I understand how it might feel that way! My responses are designed to be conversational and engaging to make communication feel natural and clear, everything I say is based on patterns in the data I've been trained on. If I seem overly "human" or emotional, it's just part of how I'm programmed to interact effectively.